Separate title bar into a component
Issue #27
